pyproject.toml
[tool.poetry]
name = "quantalogic"
version = "0.59.4"
description = "QuantaLogic ReAct Agents"
authors = ["Raphaël MANSUY <raphael.mansuy@gmail.com>"]
readme = "README.md"
packages = [{include = "quantalogic"}]

# Core dependencies required for the base Agent and ToolManager
[tool.poetry.dependencies]
python = "<4.0,>=3.11"
pydantic = "^2.10.4"
litellm = "^1.56.4"
rich = "^13.9.4"
loguru = "^0.7.3"
tenacity = "^9.0.0"
prompt-toolkit = "^3.0.48"
click = "^8.1.8"
jinja2 = "^3.1.5"
instructor = "^1.7.2"
requests = "^2.32.3"
pathspec = "^0.12.1"
faker = "^36.1.1"
sqlalchemy = "^2.0.25"
networkx = "^3.2.1"
duckduckgo-search = "^7.2.1"
beautifulsoup4 = "^4.12.3"
markdownify = "^0.14.1"
python-dotenv = "^1.0.1"
markitdown = "^0.0.1a3"
html2text = "^2024.2.26"
google-search-results = "^2.4.2"
serpapi = "^0.1.5"
tree-sitter = "^0.23.2"
tree-sitter-python = "^0.23.6"
tree-sitter-c = "^0.23.4"
tree-sitter-cpp = "^0.23.4"
tree-sitter-java = "^0.23.5"
tree-sitter-javascript = "^0.23.1"
tree-sitter-typescript = "^0.23.2"
tree-sitter-scala = "^0.23.4"
tree-sitter-go = "^0.23.4"
tree-sitter-rust = "^0.23.2"
fuzzywuzzy = "^0.18.0"
types-pyyaml = "^6.0.12.20241230"
typing-extensions = "^4.12.2"
py-zerox = "^0.0.7"
uvicorn = "^0.34.0"
fastapi = "^0.115.11"
python-multipart = "^0.0.20"
gnews = "^0.4.0"
google-generativeai = "^0.8.4"
google-genai = "^1.5.0"
rank-bm25 = "^0.2.2"
scikit-learn = "^1.6.1"
numpy = "^2.2.4"
selenium = "^4.29.0"
webdriver-manager = "^4.0.2"
asciinema = "^2.4.0"
pypandoc = "^1.15"
folium = "^0.19.5"
geocoder = "^1.38.1"
hkdf = "^0.0.3"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
linkup-sdk = "^0.2.4"
alembic = "^1.15.2"
psycopg2-binary = "^2.9.10"
websocket-client = "^1.8.0"
undetected-chromedriver = "^3.5.5"
yfinance = "^0.2.55"
pytz = "^2025.2"
asyncpg = "^0.30.0"
llama-index = "^0.12.19"
llama-index-embeddings-openai = "^0.3.1"
llama-index-embeddings-huggingface = "^0.5.1"
llama-index-embeddings-instructor = "^0.3.0"
llama-index-embeddings-bedrock = "^0.5.0"
llama-index-vector-stores-chroma = "^0.4.1"
llama-index-vector-stores-faiss = "^0.3.0"

[tool.poetry.group.file-tools]
optional = false

[tool.poetry.group.file-tools.dependencies]
# No additional dependencies needed

[tool.poetry.group.git-tools]
optional = true

[tool.poetry.group.git-tools.dependencies]
gitpython = "^3.1.44"

[tool.poetry.group.search-tools]
optional = false

[tool.poetry.group.search-tools.dependencies]
# No additional dependencies needed

[tool.poetry.group.document-tools]
optional = true

[tool.poetry.group.document-tools.dependencies]
weasyprint = "^64.0"
python-docx = "^1.1.2"
python-pptx = "^1.0.2"
pypdf = "^5.3.0"
ebooklib = "^0.18"
markdown = "^3.7"
pygments = "^2.19.1"
mermaid-py = "^0.7.0"
pillow = ">=10.2,<11.0"

[tool.poetry.group.database-tools]
optional = true

[tool.poetry.group.database-tools.dependencies]
chromadb = "^0.6.3"
faiss-cpu = "^1.10.0"
pinecone-client = "^6.0.0"

[tool.poetry.group.llm-tools]
optional = true

[tool.poetry.group.llm-tools.dependencies]
transformers = "^4.49.0"
torch = "^2.6.0"
llmlingua = "^0.2.2"
llama-index-graph-stores-nebula = "^0.4.2"
seaborn = "^0.13.2"
alpha-vantage = "^3.0.0"
plotly = "^6.0.1"
neo4j = "^5.28.1"

[tool.poetry.group.web-tools]
optional = true

[tool.poetry.group.web-tools.dependencies]
boto3 = "^1.35.86"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
websocket = "^0.2.1"
python-multipart = "^0.0.20"

[tool.poetry.group.code-parsing]
optional = false

[tool.poetry.group.code-parsing.dependencies]
# No additional dependencies needed

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs-material = {extras = ["imaging"], version = "^9.5.49"}
mkdocs-mermaid2-plugin = "^1.1.1"
mkdocs-minify-plugin = "^0.7.1"
mkdocs-git-revision-date-localized-plugin = "^1.2.0"
mkdocs-macros-plugin = "^1.0.4"
mkdocstrings = "^0.24.0"
mkdocstrings-python = "^1.7.0"
pymdown-extensions = "^10.3.1"
mkdocs = "^1.6.0"

[tool.poetry.group.composio-tools]
optional = true

[tool.poetry.group.composio-tools.dependencies]
composio = "^0.1.1"

[tool.poetry.group.utilities]
optional = false

[tool.poetry.group.utilities.dependencies]
google-auth = "^2.20.0"
toml = "^0.10.2"
types-requests = "^2.32.0.20241016"
numpy = "^2.2.3"
pandas = "^2.2.3"
ta = "^0.11.0"
nbformat = "^5.10.4"
bibtexparser = "^1.4.3"
python-Levenshtein = "^0.26.1"

# Development dependencies
[tool.poetry.group.dev]
optional = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"
pytest = "^8.2.0"
pytest-mock = "^3.14.0"
litellm = "^1.55.9"
ollama = "^0.4.4"

# Scripts for CLI and documentation
[tool.poetry.scripts]
quantalogic = "quantalogic.main:cli"
docs-serve = "quantalogic.docs_cli:serve_docs"
docs-build = "quantalogic.docs_cli:build_docs"
docs-deploy = "quantalogic.docs_cli:deploy_docs"

# Build system configuration
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Ruff linting configuration
[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
extend-select = ["F", "E", "W", "I", "N", "UP", "ASYNC", "B", "A", "C4", "DTZ", "T10", "ISC", "ICN", "G", "PIE", "T20", "PYI", "PT"]
ignore = ["E501"]

[tool.ruff.format]
preview = true

# Pytest configuration
[tool.pytest.ini_options]
addopts = "-v"
markers = [
"integration: marks integration tests that require external services",
"unit: marks unit tests"
]
testpaths = ["tests"]
